Fixed more pointer size issues

+++ vroom/r_data.c
@@ -46,6 +46,7 @@ rcsid[] = "$Id: r_data.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 #include  <alloca.h>
 #endif
 
+#include <stdint.h>
 
 #include "r_data.h"
 
@@ -639,8 +640,8 @@ void R_InitColormaps (void)
     //  256 byte align tables.
     lump = W_GetNumForName("COLORMAP"); 
     length = W_LumpLength (lump) + 255; 
-    colormaps = Z_Malloc (length, PU_STATIC, 0); 
-    colormaps = (byte *)( ((int)colormaps + 255)&~0xff); 
+    colormaps = Z_Malloc (length, PU_STATIC, 0);
+    colormaps = (byte *)( ((uintptr_t)colormaps + 255)&~0xff);
     W_ReadLump (lump,colormaps); 
 }
